pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash-iconify>=0.1.2",
    "dash-mantine-components>=2.6.0",
    "firebase>=4.0.1",
    "firebase-admin>=6.0.0,<7",
    "python-dotenv>=1.0.0",
]
